requirements.txt
discord-py
pynacl
yt-dlp
discord_components
flask